+++ packages/plugin-react-app/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.2.1
 
+- [chore] bump version of `esbuild`(0.14.14)
 - [feat] define variables of `CDN_PATH`
 - [fix] support userConfig `cssChunkNames`
 - [fix] bump version of `@builder/pack`(^0.6.0)